# src/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/user":{"get":{"operationId":"UsersController_whoAmI","parameters":[],"responses":{"200":{"description":""}}}},"/auth/signout":{"post":{"operationId":"UsersController_signOut","parameters":[],"responses":{"201":{"description":""}}}},"/auth/signup":{"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}}},"/auth/signin":{"post":{"operationId":"UsersController_signIn","parameters":[],"responses":{"201":{"description":""}}}},"/auth/{id}":{"get":{"operationId":"UsersController_findUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_removeUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}}},"/auth":{"get":{"operationId":"UsersController_findByMail","parameters":[{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/auth/user/users":{"get":{"operationId":"UsersController_findByUsername","parameters":[{"name":"user","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/reports":{"get":{"operationId":"ReportsController_getEstimate","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"ReportsController_createReport","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReportDto"}}}},"responses":{"201":{"description":""}}}},"/reports/{id}":{"get":{"operationId":"ReportsController_findReport","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ReportsController_approveReport","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveReportDto"}}}},"responses":{"200":{"description":""}}}},"/reports/reports/all":{"get":{"operationId":"ReportsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/messages":{"get":{"operationId":"MessagesController_listMessages","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"MessagesController_createMessage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"201":{"description":""}}}},"/messages/{id}":{"get":{"operationId":"MessagesController_getMessage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/logs/level":{"get":{"operationId":"LoggerController_getLogs","parameters":[],"responses":{"200":{"description":""}}}},"/logs/{newLevel}":{"put":{"operationId":"LoggerController_setLevel","parameters":[{"name":"newLevel","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Car Value App","description":"Swagger Documentation for car value app made with NestJS","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000/"}],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateReportDto":{"type":"object","properties":{}},"ApproveReportDto":{"type":"object","properties":{}},"CreateMessageDto":{"type":"object","properties":{}}}}}