# src/users/dtos/create-user.dto.ts
import { IsBoolean, IsEmail, IsOptional, IsString } from 'class-validator';

export class CreateUserDto {
  @IsEmail()
  email: string;

  @IsString()
  password: string;

  @IsString()
  @IsOptional()
  username: string;

  @IsBoolean()
  admin: boolean;
}